+++ src/terrain/voxels/soil_grass.c
@@ -6,7 +6,7 @@ ecs_entity_t spawn_block_soil_grass(ecs_world_t *world, const byte index, char* 
     zox_set_unique_name(vox, "block_soil_grass")
     // use instanced mesh prefab
     SpawnBlock data = {
-        .index = (byte) (index),
+        .index = index,
         .seed = generate_voxel_seed(index),
         .prefab = prefab_block_vox_meta,
         .prefab_texture = prefab_vox_texture,
